refactor: extract document rendering routes
